feat(organizando views): separo la logica de todas las vistas en views separados
