+++ AppCrud/views/__init__.py
@@ -1,0 +1,11 @@
+# Import all views to make them accessible from the views package
+from .inicio_views import inicio, get_job_description, enviar_correo
+from .job_views import *
+from .contacto_views import *
+from .aviso_views import *
+from .bitacora_views import *
+from .empresa_views import *
+from .usuario_views import *
+from .transacciones_views import *
+from .servidores_views import *
+from .monitoreo_views import *
